+++ docs/changelog.md
@@ -17,6 +17,9 @@ Types of changes
 
 ## Unreleased
 
+Added:
+
+- Add **Waitress** as a faster HTTP server. Werkzeug can still be used with `--http-server werkzeug` on the command line for `serve`. ([GH-422](https://github.com/martin-ueding/geo-activity-playground/issues/422))
 
 ## Version 1.26.0 — 2026-04-05